Add support `timeZoneName`
Adding the `timeZoneName="long"` or `timeZoneName="short"` attribute
adds the current timezone representation to the text visible to the
user. This is useful for situations in which the UI needs to make
explicit to the user that the time is rendered in their own local
timezone, rather than an arbitrary one determined by the server.